moved functions to controls.ts
